% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_FDR_functions.R
\name{complexCentricManualFDR}
\alias{complexCentricManualFDR}
\title{complexCentricManualFDR}
\usage{
complexCentricManualFDR(auto, manual, grid_search_list = FALSE)
}
\arguments{
\item{auto}{data.table containing complex feature finding results}

\item{manual}{data.table with manual annotations.}

\item{grid_search_list}{logical if grid search results are used}
}
\value{
List with stats - complex centric, no feature mapping performed
}
\description{
complexCentricManualFDR.
}
