% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_FDR_functions.R
\name{estimateGridSearchComplexCentricManualFDR}
\alias{estimateGridSearchComplexCentricManualFDR}
\title{estimateGridSearchComplexCentricManualFDR}
\usage{
estimateGridSearchComplexCentricManualFDR(complex_features_list,
  manual_annotation)
}
\arguments{
\item{complex_features_list}{list containing complex feature finding results for different parameter sets.}

\item{manual_annotation}{data.table with manual annotations.}
}
\value{
List with stats - complex centric, no feature mapping performed
}
\description{
estimateGridSearchComplexCentricManualFDR.
}
