% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_FDR_functions.R
\name{estimateGridSearchFeatureCentricManualFDR}
\alias{estimateGridSearchFeatureCentricManualFDR}
\title{estimateGridSearchFeatureCentricManualFDR}
\usage{
estimateGridSearchFeatureCentricManualFDR(complex_features_list,
  manual_annotation, parallelized = FALSE, n_cores = 1)
}
\arguments{
\item{complex_features_list}{list containing complex feature finding results for different parameter sets.}

\item{manual_annotation}{data.table with manual annotations.}

\item{parallelized}{logical}

\item{n_cores}{numeric number of cores to use (if parallelized == TRUE)}
}
\value{
List with stats - feature centric, all features were mapped between automated and manual
}
\description{
estimateGridSearchFeatureCentricManualFDR.
}
