% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_FDR_functions.R
\name{manualFeatureMapping}
\alias{manualFeatureMapping}
\title{manualFeatureMapping}
\usage{
manualFeatureMapping(feature, dist_cutoff, manual_annotation)
}
\arguments{
\item{feature}{feature}

\item{dist_cutoff}{dist_cutoff}

\item{manual_annotation}{manual_annotation}
}
\description{
manualFeatureMapping.
}
