% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManualComplexFeatureComparison.R
\name{plotManualComplexFeatureComparison}
\alias{plotManualComplexFeatureComparison}
\title{plotManualComplexFeatureComparison}
\usage{
plotManualComplexFeatureComparison(mapped_features, traces, manual_stats,
  calibration, complexID, TP = FALSE)
}
\arguments{
\item{mapped_features}{data.table with mapped manual and automatically detected features}

\item{traces}{protein traces object}

\item{manual_stats}{data.table with manual stats}

\item{calibration}{calibration functions from SECprofiler}

\item{complexID}{character string with complex_id}

\item{TP}{logical default TRUE}
}
\description{
plotManualComplexFeatureComparison.
}
