% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_FDR_functions.R
\name{plotManualVsDecoyFDR}
\alias{plotManualVsDecoyFDR}
\title{plotManualVsDecoyFDR}
\usage{
plotManualVsDecoyFDR(stats, FDR_cutoff = 0.1,
  colour_parameter = "completeness_cutoff", PDF = TRUE,
  name = "ManualVsDecoyFDR_plot.pdf")
}
\arguments{
\item{stats}{mergen manual and automatic stats}

\item{FDR_cutoff}{numeric default = 0.1}

\item{colour_parameter}{character string default = "completeness_cutoff"}

\item{PDF}{logical default=TRUE}

\item{name}{character string default = "ManualVsDecoyFDR_plot.pdf"}
}
\value{
List with stats - feature centric, all features were mapped between automated and manual
}
\description{
plotManualVsDecoyFDR.
}
