% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarizedManualCompleteness.R
\name{plotSummarizedManualCompleteness}
\alias{plotSummarizedManualCompleteness}
\title{plotSummarizedManualCompleteness}
\usage{
plotSummarizedManualCompleteness(hypotheses, protTraces, PDF = TRUE)
}
\arguments{
\item{hypotheses}{data.table with complex hypotheses}

\item{protTraces}{traces object of type proteins}

\item{PDF}{logical default = TRUE}
}
\description{
plotSummarizedManualCompleteness
}
